# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "HeyCreator API",
    "description": "Influencer discovery and campaign management platform API. Connects brands with creators across Instagram, TikTok, YouTube, Twitter, and Facebook.",
    "version": "1.0.0",
    "contact": {
      "name": "HeyCreator Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development"
    }
  ],
...
  "security": [
    { "BearerAuth": [] }
  ],
  "paths": {
...
    "/api/auth/update-email": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Update user email",
        "description": "Updates the email address for a user (primarily for Instagram OAuth users who need to set an email).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["uid", "email"],
                "properties": {
                  "uid": { "type": "string" },
                  "email": { "type": "string", "format": "email" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Email updated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
...
    "/api/campaigns/{id}": {
      "get": {
        "tags": ["Campaigns"],
        "summary": "Get campaign with applications",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Campaign details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "campaign": { "$ref": "#/components/schemas/Campaign" },
                        "applications": { "type": "array", "items": { "$ref": "#/components/schemas/Application" } }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": { "description": "Campaign not found" }
        }
      },
      "patch": {
        "tags": ["Campaigns"],
        "summary": "Update campaign",
        "description": "Partially updates a campaign. Brand must own the campaign.",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CampaignInput" }
            }
          }
        },
        "responses": {
          "200": { "description": "Campaign updated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "description": "Campaign not found" }
        }
      },
      "delete": {
        "tags": ["Campaigns"],
        "summary": "Delete campaign",
        "description": "Deletes a campaign. Brand must own the campaign.",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Campaign deleted", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" }, "message": { "type": "string" } } } } } },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "description": "Campaign not found" }
        }
      }
    },
...
    "/api/brands/lists": {
      "get": {
        "tags": ["Creator Lists"],
        "summary": "List creator lists",
        "description": "Fetches all creator lists for the authenticated brand.",
        "responses": {
          "200": {
            "description": "Creator lists",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/CreatorList" } },
                    "count": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Creator Lists"],
        "summary": "Create creator list",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string" },
                  "description": { "type": "string" },
                  "color": { "type": "string" },
                  "tags": { "type": "array", "items": { "type": "string" } },
                  "mentions": { "type": "array", "items": { "type": "string" } },
                  "locations": { "type": "array", "items": { "type": "string" } },
                  "influenceSize": { "type": "string" },
                  "visibility": { "type": "string", "enum": ["public", "private"] }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "List created", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" }, "data": { "$ref": "#/components/schemas/CreatorList" } } } } } },
          "400": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Firebase ID Token",
        "description": "Firebase Authentication ID token. Obtain by calling `firebase.auth().currentUser.getIdToken()`."
      },
      "UserIdHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "x-user-id",
        "description": "Development fallback: Firebase UID passed directly as header. Not for production use."
      }
    },
    "schemas": {
...
      "CampaignStatus": {
        "type": "string",
        "enum": ["DRAFT", "PUBLISHED", "ACTIVE", "IN_PROGRESS", "COMPLETED", "CANCELLED"]
      },
      "ApplicationStatus": {
        "type": "string",
        "enum": ["PENDING", "UNDER_REVIEW", "ACCEPTED", "REJECTED", "WITHDRAWN"]
      },
...
      "Campaign": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "brandId": { "type": "string" },
          "status": { "$ref": "#/components/schemas/CampaignStatus" },
          "campaignTitle": { "type": "string" },
          "description": { "type": "string" },
          "campaignObjectives": { "type": "array", "items": { "type": "string" } },
          "kpi": { "type": "string" },
          "campaignVisibility": { "type": "string", "enum": ["public", "private"] },
          "campaignStart": { "type": "string", "format": "date-time" },
          "campaignEnd": { "type": "string", "format": "date-time" },
          "campaignProduct": {
            "type": "object",
            "properties": {
              "productType": { "type": "string" },
              "productName": { "type": "string" },
              "productValue": { "type": "number" }
            }
          },
          "budget": {
            "type": "object",
            "properties": {
              "compensationModel": { "type": "string", "enum": ["fixed", "range"] },
              "currency": { "type": "string" },
              "fixedAmount": { "type": "number" },
              "minRangeAmount": { "type": "number" },
              "maxRangeAmount": { "type": "number" }
            }
          },
          "audience": {
            "type": "object",
            "properties": {
              "ageMin": { "type": "integer" },
              "ageMax": { "type": "integer" },
              "gender": { "type": "string" },
              "targetLocation": { "type": "string" },
              "minFollowers": { "type": "integer" }
            }
          },
          "tasks": {
            "type": "object",
            "properties": {
              "requiredDeliverables": { "type": "array", "items": { "type": "string" } },
              "dos": { "type": "array", "items": { "type": "string" } },
              "donts": { "type": "array", "items": { "type": "string" } }
            }
          },
          "stats": {
            "type": "object",
            "properties": {
              "views": { "type": "integer" },
              "applications": { "type": "integer" },
              "acceptedApplications": { "type": "integer" }
            }
          },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "CampaignInput": {
        "type": "object",
        "properties": {
          "campaignTitle": { "type": "string", "minLength": 3 },
          "description": { "type": "string", "minLength": 10 },
          "campaignObjectives": { "type": "array", "items": { "type": "string" } },
          "kpi": { "type": "string" },
          "isDraft": { "type": "boolean", "description": "If true, validation is relaxed" },
          "campaignVisibility": { "type": "string", "enum": ["public", "private"] },
          "campaignStart": { "type": "string", "format": "date-time" },
          "campaignEnd": { "type": "string", "format": "date-time" }
        }
      },
      "Application": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "campaignId": { "type": "string" },
          "creatorId": { "type": "string" },
          "status": { "$ref": "#/components/schemas/ApplicationStatus" },
          "campaignTitle": { "type": "string" },
          "coverLetter": { "type": "string" },
          "proposedRate": { "type": "number" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "CreatorList": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "color": { "type": "string" },
          "creatorCount": { "type": "integer" },
          "tags": { "type": "array", "items": { "type": "string" } },
          "mentions": { "type": "array", "items": { "type": "string" } },
          "locations": { "type": "array", "items": { "type": "string" } },
          "influenceSize": { "type": "string" },
          "visibility": { "type": "string", "enum": ["public", "private"] },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "data": { "type": "object" },
          "count": { "type": "integer" }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "error": { "type": "string" },
          "code": { "type": "string" },
          "details": { "type": "string" }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request — validation error or missing fields",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
...
      "Forbidden": {
        "description": "Forbidden — insufficient permissions (wrong role)",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "ServerError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      }
    }
  }
}